refactor: create package runner
